Lab3

+++ src/domain/InstructionElement.java
@@ -3,9 +3,9 @@ package domain;
 public class InstructionElement {
     public InstructionTypeEnum instruction;
     public RegisterElement registerA;
-    public Integer valueA;
+    public Byte valueA;
     public RegisterElement registerB;
-    public Integer valueB;
+    public Byte valueB;
     public RegisterElement destiny;
 
     public InstructionElement(InstructionTypeEnum instruction, RegisterElement paramA, RegisterElement paramB, RegisterElement destiny) {
@@ -17,7 +17,7 @@ public class InstructionElement {
         this.destiny = destiny;
     }
 
-    public InstructionElement(InstructionTypeEnum instruction, RegisterElement paramA, Integer paramB, RegisterElement destiny) {
+    public InstructionElement(InstructionTypeEnum instruction, RegisterElement paramA, Byte paramB, RegisterElement destiny) {
         this.instruction = instruction;
         registerA = paramA;
         valueA = paramA.value;
@@ -26,7 +26,7 @@ public class InstructionElement {
         this.destiny = destiny;
     }
 
-    public InstructionElement(InstructionTypeEnum instruction, Integer paramA, RegisterElement paramB, RegisterElement destiny) {
+    public InstructionElement(InstructionTypeEnum instruction, Byte paramA, RegisterElement paramB, RegisterElement destiny) {
         this.instruction = instruction;
         registerA = null;
         valueA = paramA;
@@ -35,7 +35,7 @@ public class InstructionElement {
         this.destiny = destiny;
     }
 
-    public InstructionElement(InstructionTypeEnum instruction, Integer paramA, Integer paramB, RegisterElement destiny) {
+    public InstructionElement(InstructionTypeEnum instruction, Byte paramA, Byte paramB, RegisterElement destiny) {
         this.instruction = instruction;
         registerA = null;
         valueA = paramA;
