+++ src/domain/RegisterElement.java
@@ -2,7 +2,7 @@ package domain;
 
 public class RegisterElement {
     public boolean busy;
-    public Integer value;
+    public Byte value;
 
     public RegisterElement() {
         busy = false;
